Revert "DAOS-3115 vos: Add new tests for punch model change (#1351)"
This reverts commit b420b7245fcd6fd0222496dd4e1ac1a29be8f499.
