+++ src/vos/tests/SConscript
@@ -15,8 +15,7 @@ def scons():
 
     vos_test_src = ['vos_tests.c', 'vts_io.c', 'vts_pool.c', 'vts_container.c',
                     denv.Object("vts_common.c"), 'vts_aggregate.c', 'vts_dtx.c',
-                    'vts_gc.c', 'csum_extent_tests.c', 'vts_ilog.c',
-                    'vts_array.c', 'vts_pm.c']
+                    'vts_gc.c', 'csum_extent_tests.c', 'vts_ilog.c']
     vos_tests = daos_build.program(denv, 'vos_tests', vos_test_src,
                                    LIBS=libraries)
     denv.AppendUnique(CPPPATH=["../../common/tests"])
