+++ src/vos/tests/vos_tests.c
@@ -55,7 +55,6 @@ print_usage()
 	print_message("vos_tests -A|--all_tests\n");
 	print_message("vos_tests -f|--filter <filter>\n");
 	print_message("vos_tests -e|--exclude <filter>\n");
-	print_message("vos_tests -m|--punch-model-tests\n");
 	print_message("vos_tests -h|--help\n");
 	print_message("Default <vos_tests> runs all tests\n");
 }
@@ -82,7 +81,6 @@ run_all_tests(int keys, bool nest_iterators)
 	int	i;
 	int	j;
 
-	failed += run_pm_tests();
 	failed += run_pool_test();
 	failed += run_co_test();
 	for (i = 0; dkey_feats[i] >= 0; i++) {
@@ -109,7 +107,7 @@ main(int argc, char **argv)
 	int	ofeats;
 	int	keys;
 	bool	nest_iterators = false;
-	const char *short_options = "apcdglni:mXA:hf:e:";
+	const char *short_options = "apcdglni:XA:hf:e:";
 	static struct option long_options[] = {
 		{"all_tests",		required_argument, 0, 'A'},
 		{"pool_tests",		no_argument, 0, 'p'},
@@ -119,7 +117,6 @@ main(int argc, char **argv)
 		{"nest_iterators",	no_argument, 0, 'n'},
 		{"aggregate_tests",	no_argument, 0, 'a'},
 		{"dtx_tests",		no_argument, 0, 'X'},
-		{"punch_model_tests",	no_argument, 0, 'm'},
 		{"garbage_collector",	no_argument, 0, 'g'},
 		{"ilog_tests",		no_argument, 0, 'l'},
 		{"help",		no_argument, 0, 'h'},
@@ -204,10 +201,6 @@ main(int argc, char **argv)
 			break;
 		case 'X':
 			nr_failed += run_dtx_tests();
-			test_run = true;
-			break;
-		case 'm':
-			nr_failed += run_pm_tests();
 			test_run = true;
 			break;
 		case 'A':
