+++ src/vos/tests/vts_common.h
@@ -126,7 +126,6 @@ int run_discard_tests(void);
 int run_aggregate_tests(bool slow);
 int run_dtx_tests(void);
 int run_gc_tests(void);
-int run_pm_tests(void);
 int run_io_test(daos_ofeat_t feats, int keys, bool nest_iterators);
 
 int run_ilog_tests(void);
